Add files via upload
1. Video<-> Images: Write a program to convert a given video to its constituent images. Your output should be
in a specified folder. Write another program that will merge a set of images in a folder into a single video. You
should be able to control the frame rate in the video that is created.
2. Capturing Images: Learn how to capture frames from a webcam connected to your computer and save them as
images in a folder. You may use either the built-in camera of your laptop or an external one connected through
USB. You should also be able to display the frames (the video) on the screen while capturing.